// src/app/models/driver.model.ts
export class Driver {
   id: number;
   foto: string;
   familia: string;
   name: string;
   name2: string;
   payment: any;
   data: number;
   sex: boolean;
   paspseria: string;
   paspsnumber: number;
   paspsdata: number;
   paspname: string;
   paspregistr: string;
   pasplive: string;
   phone: number;
}